clear session response logic
