+++ app/controllers/api/v1/protected_application_controller.rb
@@ -2,6 +2,7 @@ module Api
   module V1
     class ProtectedApplicationController < ApplicationController
       before_action :authenticate_request
+
       include ActionController::Cookies
 
       private
