+++ app/serializers/user_serializer.rb
@@ -1,4 +1,4 @@
 class UserSerializer
   include JSONAPI::Serializer
-  attributes :id, :email, :username
+  attributes :email, :username
 end
